django-pip/requirements.txt
Django==4.1.7
gunicorn==21.2.0
djangorestframework==3.14.0
django-rest-passwordreset==1.3.0
django-rest-registration==0.8.3
django-cors-headers==3.14.0
djangorestframework-simplejwt==4.8.0
mysqlclient==2.1.1


